docker_mount/project/code/retrieval/requirements.txt
torch
torchvision
transformers
pillow
rasterio
mercantile
numpy
qdrant-client
pydantic
pyyaml
tqdm
pandas
requests
sentencepiece
pyarrow
